pyproject.toml
[project]
name = "btc-dollar-cost"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Noboru Katayama", email = "katayama@rs.tus.ac.jp" }
]
dependencies = [
    "python-dotenv>=1.0.1",
    "pybitflyer>=0.1.9",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/btc_dollar_cost"]
